requirements.txt
langdetect
iso639-lang
rake_nltk